Try to use trait object suggestion from SO
