+++ dev/playground/src/main.rs
@@ -1,77 +1,45 @@
-
-// https://stackoverflow.com/questions/49711479/how-can-i-create-hashable-trait-objects-trait-objects-with-generic-method-para
-
-use std::collections::HashMap;
+use std::hash::{Hash, Hasher, BuildHasher};
 use std::collections::hash_map::RandomState;
-use std::hash::Hasher;
-use std::hash::BuildHasher;
-use std::hash::Hash;
 
-#[macro_use]
-extern crate lazy_static;
+#[derive(Hash)]
+struct Foo(i32);
+
+#[derive(Hash)]
+struct Bar(String);
 
 trait MyTrait {
-    fn to_text(&self) -> String;
+    fn my_hash(&self, h: &mut Hasher);
 }
 
-#[derive(Hash, PartialEq, Eq, Debug)]
-struct Alpha {
-    val: String,
-}
-
-#[derive(Hash, PartialEq, Eq, Debug)]
-struct Beta {
-    nr: i32,
-    f: u8,
-}
-
-impl MyTrait for Alpha {
-    fn to_text(&self) -> String {
-        self.val.to_owned()
+impl MyTrait for Foo {
+    fn my_hash(&self, mut h: &mut Hasher) {
+        self.hash(&mut h)
     }
 }
 
-impl MyTrait for Beta {
-    fn to_text(&self) -> String {
-        format!("{}, {}", self.nr, self.f)
+impl MyTrait for Bar {
+    fn my_hash(&self, mut h: &mut Hasher) {
+        self.hash(&mut h)
     }
 }
 
-#[derive(Hash, PartialEq, Eq, Debug)]
-enum MyEnum {
-    A(Alpha),
-    B(Beta),
-}
-
-impl MyTrait for MyEnum {
-    fn to_text(&self) -> String {
-        match self {
-            &MyEnum::A(ref alpha) => alpha.to_text(),
-            &MyEnum::B(ref beta) => beta.to_text(),
-        }
+impl Hash for MyTrait {
+    fn hash<H: Hasher>(&self, hasher: &mut H) {
+        self.my_hash(hasher)
     }
 }
 
-lazy_static! {
-    static ref RANDSTATE: RandomState = RandomState::new();
-}
-
-fn get_test_hash(x: &MyEnum) -> u64 {
-   	let mut hasher = RANDSTATE.build_hasher();
+fn get_test_hash(x: &MyTrait) -> u64 {
+    let randstate = RandomState::new();
+   	let mut hasher = randstate.build_hasher();
     x.hash(&mut hasher);
     hasher.finish()
 }
 
 fn main() {
-    let a1: MyEnum = MyEnum::A(Alpha { val: "Hello World".to_owned() });
-    let a2: MyEnum = MyEnum::A(Alpha { val: "Bye World".to_owned() });
-    let a3: MyEnum = MyEnum::A(Alpha { val: "Bye World".to_owned() });
-    let b: MyEnum = MyEnum::B(Beta { nr: 8, f: 2 });
-    let mut m = HashMap::new();
-    println!("{:?} {:?}", a1.to_text(), b.to_text());
-    println!("{:?} {:?} {:?}", get_test_hash(&a1), get_test_hash(&a2), get_test_hash(&a3));
-    m.insert(a1, 0);
-    m.insert(a2, 0);
-    m.insert(b, 0);
-    println!("{:?}", m);
+    let foo = Foo(42);
+    let bar = Bar("answer".into());
+
+    println!("{:?}", get_test_hash(&foo));
+    println!("{:?}", get_test_hash(&bar));
 }
